Bug fixes
-Fixed a bug where the enemy would not change color after color scheme
change
- Fixed a bug where an enemy killing you with bullets in the
options/menu screen would make the preview stop

+++ On the Line/On the Line/Obstacles.cs
@@ -133,7 +133,7 @@ namespace On_the_Line
                         Show = false;
                     }
                 }
-                if ((OnTheLine.screen == (int)Screen.GameScreen && (Position.Y > 930 && Color != OnTheLine.TextColor && MaxMove == 0)) || _gateway)
+                if ((OnTheLine.screen == Screen.GameScreen && (Position.Y > 930 && Color != OnTheLine.TextColor && MaxMove == 0)) || _gateway)
                 {
                     Show = false;
                 }
